Refactor ClassNames tests
